+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * 【http】         HttpRequest增加setChunkedStreamingMode方法（issue#525@Github）
 * 【setting】     SettingLoader支持自定义分隔符
 * 【http】         Content-Type添加默认值（issue#I11YHI@Gitee）
+* 【socket】     增加Closeable接口（issue#532@Github）
 
 ### Bug修复
 * 【core】         修复NetUtil.getUsableLocalPort问题（pr#69@Gitee）